requirements.txt
anthropic>=0.40.0
pymupdf>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
arxiv>=2.1.0
rich>=13.0.0
